requirements.txt
langchain
langchain_openai
load_dotenv
langchain_community
tavily-python

fastapi
uvicorn
openai
websockets
sounddevice
soundfile
numpy
pyttsx3
SpeechRecognition
fastrtc[stt]
fastrtc-whisper-cpp
elevenlabs
websocket-client
pyaudio